[TASK] Loosen match for URL to ignore query string and trailing slash
Loosens the match logic to ignore traling slashes and the query string. If a query string is provided it will be taken into account, however it will match a redirect url without it and append the query string when redirecting. This can be used for pages that differ on the query string, but all goes to the same page.